extraction_scripts/macOS/apple_vision_ocr/requirements.txt
pyobjc-framework-Vision==10.1
pyobjc-framework-Quartz==10.1
tqdm==4.66.2 